Lots of bugs fixed, build 21000
Many errors have been fixed, but not all. Initial work on Minecraft 1.10
compatibility.
WARNING: Because I downloaded this plugin from Bukkit, and decompiled
the plugin to begin my work on MyZ, the comments were deleted
unexpectedly. The official developer of MyZ will need to add the
comments back if he has the time. Sorry about that.